+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+- allow deriving a rectangular boundary from hole polygons when no explicit boundary is supplied
+
 
 2.7.2 (2024-04-10)
 -------------------